Add serializer for a `PartData`
