+++ core/mastodon/src/test/java/br/com/orcinus/orca/core/mastodon/network/request/headers/form/item/Assert.extensions.kt
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.mastodon.network.request.headers.form
+package br.com.orcinus.orca.core.mastodon.network.request.headers.form.item
 
 import assertk.Assert
 import assertk.assertions.hasSameContentAs
